requirements.txt
feedparser
requests
beautifulsoup4
python-dateutil
pandas
pyyaml
lxml
tqdm